refactor: move ArtilleryWorker definition into its own module
