+++ lib/launch-local.js
@@ -3,19 +3,18 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const divideWork = require('./dist');
-const { SSMS, normalizeTs } = require('../core/lib/ssms');
+const { SSMS } = require('../core/lib/ssms');
 const { loadPlugins } = require('./load-plugins');
-const awaitOnEE = require('./util/await-on-ee');
-const sleep = require('./util/sleep');
+
+const { ArtilleryWorker } = require('./artillery-worker-local');
+
 const EventEmitter = require('eventemitter3');
 const debug = require('debug')('core');
 
-const { Worker } = require('worker_threads');
-const path = require('path');
 const os = require('os');
 const p = require('util').promisify;
 
-async function createRunner(script, payload, opts) {
+async function createLauncher(script, payload, opts) {
   return new Launcher(script, payload, opts);
 }
 
@@ -159,6 +158,8 @@ class Launcher {
     }
   }
 
+  // TODO: Workers to be launched by LaunchPlatform.createWorker() rather
+  // than individually
   async run() {
     await this.initPlugins();
 
@@ -298,116 +299,4 @@ class Launcher {
   }
 }
 
-const STATES = {
-  initializing: 1,
-  online: 2,
-  preparing: 3,
-  readyWaiting: 4,
-  running: 5,
-  unknown: 6,
-  stoppedError: 7,
-  completed: 8,
-  stoppedEarly: 9,
-  stoppedFailed: 10,
-  timedout: 11,
-};
-
-class ArtilleryWorker {
-  constructor(opts) {
-    this.opts = opts;
-    this.events = new EventEmitter(); // events for consumers of this object
-    this.workerEvents = new EventEmitter(); // turn events delivered via 'message' events into their own messages
-  }
-
-  async init(_opts) {
-    this.state = STATES.initializing;
-
-    this.worker = new Worker(path.join(__dirname, 'worker.js'));
-    this.workerId = this.worker.threadId;
-    this.worker.on('error', this.onError.bind(this));
-    // TODO:
-    this.worker.on('exit', (exitCode) => {
-      this.events.emit('exit', exitCode);
-    });
-    this.worker.on('messageerror', (err) => {
-
-    });
-
-    // TODO: Expose performance metrics via getHeapSnapshot() and performance object.
-
-    await awaitOnEE(this.worker, 'online', 10);
-
-    // Relay messages onto the real event emitter:
-    this.worker.on('message', (message) => {
-      switch(message.event) {
-      case 'workerError':
-        this.events.emit('workerError', message);
-        this.workerEvents.emit('workerError', message);
-        break;
-      case 'phaseStarted':
-        this.events.emit('phaseStarted', message);
-        this.workerEvents.emit('phaseStarted', message);
-        break;
-      case 'phaseCompleted':
-        this.events.emit('phaseCompleted', message);
-        this.workerEvents.emit('phaseCompleted', message);
-        break;
-      case 'stats':
-        this.events.emit('stats', message);
-        this.workerEvents.emit('stats', message);
-        break;
-      case 'done':
-        this.events.emit('done', message);
-        this.workerEvents.emit('done', message);
-        break;
-      case 'running':
-        this.events.emit('running', message);
-        this.workerEvents.emit('running', message);
-        break
-      case 'readyWaiting':
-        this.events.emit('readyWaiting', message);
-        this.workerEvents.emit('readyWaiting', message);
-        break;
-      default:
-        global.artillery.log(`Unknown message from worker ${message}`, 'error');
-      }
-    });
-
-    this.state = STATES.online;
-  }
-
-  async prepare(opts) {
-    this.state = STATES.preparing;
-
-    const { script, payload, options } = opts;
-    this.worker.postMessage({
-      command: 'prepare',
-      opts: { script, payload, options }
-    });
-
-    await awaitOnEE(this.workerEvents, 'readyWaiting', 50);
-    this.state = STATES.readyWaiting;
-  }
-
-  async run(opts) {
-    this.worker.postMessage({
-      command: 'run',
-      opts,
-    });
-
-    await awaitOnEE(this.workerEvents, 'running', 50);
-    this.state = STATES.running;
-  }
-
-  async stop() {
-    this.worker.postMessage({ command: 'stop' });
-  }
-
-  onError(err) {
-    // TODO: set state, clean up
-    this.events.emit('error', err);
-    console.log('worker error, id:', this.workerId, err);
-  }
-}
-
-module.exports = createRunner;
+module.exports = createLauncher;
